blogosphere/blogs/requirements.txt
Django==1.11.10
django-admin-sortable==2.1.5
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
django-comments-xtd==2.1.0
django-contrib-comments==1.8.0
django-js-asset==1.0.0
django-modeltranslation==0.12.2
django-mptt==0.9.0
django-mptt-admin==0.5.1
django-tagging==0.4.6
django-versatileimagefield==1.9
djangorestframework==3.8.2
docutils==0.14
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.4
six==1.11.0
